pipecat-service/requirements.txt
pipecat-ai[deepgram,anthropic,websocket]
python-dotenv
httpx
fastapi
python-multipart
uvicorn[standard]
onnxruntime
twilio
aiohttp
